*: simplify hashicorp/raft-wal implementation
This commit is a large refactoring to remove all references to raft in the WAL
and reduce the import surface area. Some packages (e.g. logging & metrics) have
been replaced by what is used at Polar Signals.

Some major changes:
- DeleteRange has been renamed to TruncateFront and TruncateBack to truncate
the WAL, which are the only two features we need (DeleteRange also didn't allow
deleting an arbitrary range, it had to be a prefix or suffix).
- Codecs have been removed. The responsibility of encoding/decoding entries
are left up to the calling library. The codec version is still encoded in
segment files in case we want to add that back in.
- MetaDB only stores state. The previous implementation also allowed arbitrary
puts/gets but this was not necessary for our use case.
- Previously, a caller had to append raft.Log entries which were then converted
to types.LogEntry. This commit makes it so that callers can write a batch of
[]types.LogEntry directly.