*: read directly into LogEntry in GetLog
Previously, the wal code maintained a fragmented sync.Pool for PooledBuffers
that was not effective. We would therefore allocate a buffer for each GetLog
call, then copy into the LogEntry (with a possible reallocation of this second
buffer to make enough space) the caller passed in.

This commit directly plumbs down this LogEntry to the readFrame code, therefore
giving callers the option to reuse the same buffer on each call to GetLog. This
can greatly help reduce allocations (both objects and bytes).