*: allow building wal with GOARCH=wasm GOOS=wasip1
This required abstracting some unix-only functionality and implementing a
non-bolt meta store. The wasm meta store is purposefully simple and not
production ready since we just need something to run tests with. Eventually,
we should remove bolt and use a simpler production-ready metastore since the
functionality Bolt was used for in the original repo is something we don't need
and have actually eliminated (arbitrary kv get/sets).